Small file movements, db auth data moved to env
