+++ src/main/java/database/DAO.java
@@ -1,7 +1,8 @@
 package database;
 
 
-import mainBot.User;
+import database.models.Connection;
+import database.models.User;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.query.Query;
