+++ src/main/java/database/models/Connection.java
@@ -1,4 +1,4 @@
-package database;
+package database.models;
 
 import jakarta.persistence.*;
