+++ src/main/resources/hibernate.cfg.xml
@@ -5,8 +5,5 @@
     <session-factory>
         <property name = "hibernate.dialect">org.hibernate.dialect.PostgreSQLDialect</property>
         <property name = "hibernate.connection.driver_class">org.postgresql.Driver</property>
-        <property name = "hibernate.connection.url">jdbc:postgresql://localhost:5432/mechmatchdb</property>
-        <property name = "hibernate.connection.username">postgres</property>
-        <property name = "hibernate.connection.password">postgrespswd</property>
     </session-factory>
 </hibernate-configuration>
